requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
python-docx>=0.8.11
docxcompose>=1.3.3
lxml>=4.9.0 